fix parse.awk returning only first match
That was because we deleted the word from the list when it was
found. This is not very cool, especially taking in account fuzzy
matching, which undoubtedly will find more than one match.

Also we fixed silly Words array filling, which prevented us from using
'in' operator. So we switched to storing useful info in keys, since we
don't even use indexes anywhere.

Reporting of words that can't be found was also adjusted accordingly.